+++ dalek/single-panel.js
@@ -12,6 +12,17 @@ module.exports = {
     .done();
   },
 
+  'should be able to trigger with event': function(test) {
+    test.open('http://localhost:3000/home')
+    .wait(500)
+    .click('[data-ut="expander1"] .handle')
+      .assert.visible('[data-ut="expander1"] .content', 'single panel visible')
+    .moveTo('[data-ut="event-expander2"]')
+      .assert.visible('[data-ut="expander2"] .content', 'new single panel visible')
+      .assert.notVisible('[data-ut="expander1"] .content', 'old single panel not visible')
+    .done();
+  },
+
   'should keeping the single panel open in clicking inside the content of the single panel': function(test) {
     test.open('http://localhost:3000/home')
     .wait(500)