+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Each change log are assumed to be accompanied by associated unit test and documentation updates.
 
+## master
+
+- added `NagSinglePanel[]/show` event to be able to trigger single panel programmatically
+
 ## 0.5.0
 
 - initial release with change log file